requirements.txt
https://github.com/KurimuzonAkuma/pyrogram/archive/d1d9989a69409ae4f71fc38ee05c53caaa8e7f4b.zip
python_dotenv==1.0.1
tgcrypto==1.2.5
